requirements.txt
# Core dependencies
python-whois==0.8.0
dnspython==2.4.2
requests==2.31.0
pandas==2.1.4
neo4j==5.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Orchestration
prefect==2.14.11

# Web framework for API/dashboard
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0

# Data processing
python-dateutil==2.8.2

# Tech stack detection (free, comprehensive)
# python-wappalyzer uses Wappalyzer's detection patterns (100+ CMS, 1000+ technologies)
python-wappalyzer==0.3.1
setuptools>=80.0.0  # Required for python-wappalyzer

# Testing
pytest==7.4.3
pytest-cov==4.1.0

